Updating rule, rule generator, and adding tests
The rule is able to be stored as a function in a JS file and then be read in as
a string with some minor parsing and token replacement in order to have it ready
to go for installation. This allows us to edit the code with full syntax
highlighting and static analysis since it does not live in a template string.